+++ shell/platform/android/external_view_embedder/external_view_embedder.cc
@@ -22,7 +22,7 @@ AndroidExternalViewEmbedder::AndroidExternalViewEmbedder(
 
 // |ExternalViewEmbedder|
 void AndroidExternalViewEmbedder::PrerollCompositeEmbeddedView(
-    int view_id,
+    int64_t view_id,
     std::unique_ptr<EmbeddedViewParams> params) {
   TRACE_EVENT0("flutter",
                "AndroidExternalViewEmbedder::PrerollCompositeEmbeddedView");
@@ -47,7 +47,7 @@ void AndroidExternalViewEmbedder::PrerollCompositeEmbeddedView(
 
 // |ExternalViewEmbedder|
 EmbedderPaintContext AndroidExternalViewEmbedder::CompositeEmbeddedView(
-    int view_id) {
+    int64_t view_id) {
   if (slices_.count(view_id) == 1) {
     return {slices_.at(view_id)->canvas(), slices_.at(view_id)->builder()};
   }
@@ -79,7 +79,7 @@ AndroidExternalViewEmbedder::GetCurrentBuilders() {
   return builders;
 }
 
-SkRect AndroidExternalViewEmbedder::GetViewRect(int view_id) const {
+SkRect AndroidExternalViewEmbedder::GetViewRect(int64_t view_id) const {
   const EmbeddedViewParams& params = view_params_.at(view_id);
   // TODO(egarciad): The rect should be computed from the mutator stack.
   // (Clipping is missing)
