+++ shell/platform/darwin/ios/framework/Source/FlutterPlatformViews_Internal.h
@@ -200,7 +200,7 @@ class FlutterPlatformViewsController {
   // Also reverts the composition_order_ to its original state at the beginning of the frame.
   void CancelFrame();
 
-  void PrerollCompositeEmbeddedView(int view_id,
+  void PrerollCompositeEmbeddedView(int64_t view_id,
                                     std::unique_ptr<flutter::EmbeddedViewParams> params);
 
   // Returns the `FlutterPlatformView`'s `view` object associated with the view_id.
@@ -208,7 +208,7 @@ class FlutterPlatformViewsController {
   // If the `FlutterPlatformViewsController` does not contain any `FlutterPlatformView` object or
   // a `FlutterPlatformView` object asscociated with the view_id cannot be found, the method
   // returns nil.
-  UIView* GetPlatformViewByID(int view_id);
+  UIView* GetPlatformViewByID(int64_t view_id);
 
   PostPrerollResult PostPrerollAction(
       const fml::RefPtr<fml::RasterThreadMerger>& raster_thread_merger);
@@ -220,11 +220,11 @@ class FlutterPlatformViewsController {
 
   std::vector<DisplayListBuilder*> GetCurrentBuilders();
 
-  EmbedderPaintContext CompositeEmbeddedView(int view_id);
+  EmbedderPaintContext CompositeEmbeddedView(int64_t view_id);
 
   // The rect of the platform view at index view_id. This rect has been translated into the
   // host view coordinate system. Units are device screen pixels.
-  SkRect GetPlatformViewRect(int view_id);
+  SkRect GetPlatformViewRect(int64_t view_id);
 
   // Discards all platform views instances and auxiliary resources.
   void Reset();
@@ -289,7 +289,7 @@ class FlutterPlatformViewsController {
   void ApplyMutators(const MutatorsStack& mutators_stack,
                      UIView* embedded_view,
                      const SkRect& bounding_rect);
-  void CompositeWithParams(int view_id, const EmbeddedViewParams& params);
+  void CompositeWithParams(int64_t view_id, const EmbeddedViewParams& params);
 
   // Allocates a new FlutterPlatformViewLayer if needed, draws the pixels within the rect from
   // the picture on the layer's canvas.
