+++ shell/platform/fuchsia/flutter/gfx_external_view_embedder.cc
@@ -170,7 +170,7 @@ GfxExternalViewEmbedder::GetCurrentBuilders() {
 }
 
 void GfxExternalViewEmbedder::PrerollCompositeEmbeddedView(
-    int view_id,
+    int64_t view_id,
     std::unique_ptr<flutter::EmbeddedViewParams> params) {
   zx_handle_t handle = static_cast<zx_handle_t>(view_id);
   FML_CHECK(frame_layers_.count(handle) == 0);
@@ -182,7 +182,7 @@ void GfxExternalViewEmbedder::PrerollCompositeEmbeddedView(
 }
 
 flutter::EmbedderPaintContext GfxExternalViewEmbedder::CompositeEmbeddedView(
-    int view_id) {
+    int64_t view_id) {
   zx_handle_t handle = static_cast<zx_handle_t>(view_id);
   auto found = frame_layers_.find(handle);
   FML_CHECK(found != frame_layers_.end());
