+++ shell/testing/tester_main.cc
@@ -55,7 +55,7 @@ class TesterExternalViewEmbedder : public ExternalViewEmbedder {
 
   // |ExternalViewEmbedder|
   void PrerollCompositeEmbeddedView(
-      int view_id,
+      int64_t view_id,
       std::unique_ptr<EmbeddedViewParams> params) override {}
 
   // |ExternalViewEmbedder|
@@ -65,7 +65,7 @@ class TesterExternalViewEmbedder : public ExternalViewEmbedder {
   std::vector<DisplayListBuilder*> GetCurrentBuilders() override { return {}; }
 
   // |ExternalViewEmbedder|
-  EmbedderPaintContext CompositeEmbeddedView(int view_id) override {
+  EmbedderPaintContext CompositeEmbeddedView(int64_t view_id) override {
     return {&canvas_, nullptr};
   }
